Random forest classification machine learning
In social media platforms, we can see various kinds of advertisements with respect to users'
perspectives over the internet. Our model represents the analysis of user purchases over social
media from advertisements. The dataset lists customer’s ages, salaries, and whether
advertisements persuaded them to make a purchase. We have two features and a target variable
that processes and analysis using Random Forest classification logic. We tested, trained, and
visualized the data from the desired dataset. In this project, we have introduced a user interface
to automate the entire process which enhanced the implementation in the Jupiter notebook